Apr19 (#103)
* added VERY basic index with pagination of current user expenses

* curreency now converted in ExpenseService inside method, Expense submission form has more fields with DTO2, decided to comment out RecieptData classes because they will be field within expense

* Testing Controllers WIP

* drop down on expense, added bean for currency, did other small things

* notes for aga

* pushing before major reorganization of controllers for expense

* small updates

* Manager actions (#89)

* added VERY basic index with pagination of current user expenses

* added capability for managers to view their expenses approval queue

* fixed relations in our database, added buttons to manager index

* fixed minor pagination issues, and combined index & expense_management controller methods into one

* removed unnecessary maincontroller method for expense_management

* Front dashboard (#90)

* added VERY basic index with pagination of current user expenses

* checking the if the upload works

* #11 nv and footer basics

* #70 implementing uploading reeipts into the expense form


* Testing Controllers WIP (#91)


* Expense Methods

* currently working just fine

* Manager actions modal (#93)

* currently working just fine

* I think its working

* solved the bug with appending to the link, need to fix the other buttons

* Html receipt (#94)

* template submission and expense service changes

* expense form submission now persists an expense entity


* #97 storing the images first try. adding methods to expense related classes

* #97 no more storing , just showing the file

* Creating more methods

* #97 cleaning up the unused imports

* more stuff

* more stuff 2

* notification button basic html

* basic html button notify

* more stuff 3

* navbar, exp approved, other small things

* Update expense (#98)

* modals added to each manager action button, modal moved to separate th fragment

* almost done with expensecontroller handling

* manager actions are complete in controller and service, data is persisted correctly as well

* fixed receipt display in the table of our index

* removed unnecessary controller method

* update expense form added and working

* added footer, checkbox, small changes

+++ src/main/java/wobbly/pigeons/expensemanager/util/ConverterRestClient.java
@@ -25,7 +25,8 @@ public class ConverterRestClient {
 
     public Long getConversionAmount (String baseCode, String targetCode, double amount){
 
-        return (Double) webClient
+     //   return (Double) webClient
+        return (Long) webClient
                 .get()
                 // .uri("/pair/{base_code}/target_code/{amount}")
                 .uri (uriBuilder -> uriBuilder
